+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Fix `Encoder[GeometryCollection]` including subclasses of GeometryCollection twice in the json
+   (MultiPolygon, Multipoint,MultiLinestring) [#3167](https://github.com/locationtech/geotrellis/issues/3167)
+
 ## [3.5.0] - 2020-08-18
 
 ### Added